refactor: move LoginController to their own production file
