+++ src/presentation/controller/index.ts
@@ -1,1 +1,2 @@
 export * from "./signUp-controller"
+export * from "./login-controller"
